requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
Deprecated==1.3.1
distro==1.9.0
fastapi==0.127.0
fastembed==0.7.4
filelock==3.20.1
flatbuffers==25.12.19
fsspec==2025.12.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.5
langchain-openai==1.1.6
langsmith==0.5.0
limits==5.6.0
loguru==0.7.3
mmh3==5.2.0
mpmath==1.3.0
numpy==2.4.0
onnxruntime==1.23.2
openai==2.14.0
orjson==3.11.5
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
prometheus_client==0.23.1
protobuf==6.33.2
py_rust_stemmers==0.1.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sentencepiece==0.2.1
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
zstandard==0.25.0
